refactor: change type file naming
