+++ src/components/bar-chart/bar-chart.tsx
@@ -1,7 +1,7 @@
 import React, { memo } from "react";
 import { Chart } from "react-google-charts";
+import { Category } from "../../api/expenses.types";
 import { Expense } from "../../store/expenseSlice";
-import { Category } from "../../api/expenses-api";
 
 type BarChartData = {
   categories: Category[];
